+++ data-raw/maya-rda/Nichia/dimming-COB/make-CC-dimming.R
@@ -45,11 +45,11 @@ for (s in names(COB_dimming.mspct)) {
                         "WithLEDiL F15559_MIRELLA-G2-M reflector 25 degrees")
   what.measured <- paste("10W Optisolis COB LED driven at ",  formatted.conditions[s])
   temp.spct <- COB_dimming.mspct[[s]]
-  temp.spct <- thin_wl(temp.spct)
+  temp.spct <- thin_wl(temp.spct, max.wl.step = 5, max.slope.delta = 0.0005)
   temp.spct <- trim_wl(temp.spct, c(330, 900), fill = 0)
   setHowMeasured(temp.spct, how.measured)
   setWhatMeasured(temp.spct, what.measured)
-  comment(temp.spct) <- "Comparison of reflectors at close range (159 mm) rather than design distance. This is most likely why highest irradiance is not observed with the narrowest beam reflector, but with a wider one."
+  comment(temp.spct) <- comment.text
   trimInstrDesc(temp.spct)
   trimInstrSettings(temp.spct)
   print(str(get_attributes(temp.spct)))
@@ -65,11 +65,27 @@ names(COB_dimming.mspct) <- gsub("\\.*3[0-9]\\.[0-9][0-9]*[vV]$", "", names(COB_
 
 q_irrad(COB_dimming.mspct, scale.factor = 1e6)
 
-autoplot(COB_dimming.mspct)
+print(autoplot(COB_dimming.mspct))
 summary(COB_dimming.mspct)
 cat(comment(COB_dimming.mspct))
-irrads <- q_irrad(COB_dimming.mspct, scale.factor = 1e6)
-irrads$Q_Total_rel <- irrads$Q_Total / max(irrads$Q_Total) * 100
-irrads
 
-save(COB_dimming.mspct, file = "./data/COB-dimming-mspct.rda")
+photons.tb <- q_irrad(COB_dimming.mspct, scale.factor = 1e6)
+energy.tb <- e_irrad(COB_dimming.mspct, scale.factor = 1)
+
+COB_dimming.tb <- full_join(photons.tb, energy.tb)
+COB_dimming.tb$Q_Total_rel <- COB_dimming.tb$Q_Total / max(COB_dimming.tb$Q_Total) * 100
+COB_dimming.tb$E_Total_rel <- COB_dimming.tb$E_Total / max(COB_dimming.tb$E_Total) * 100
+
+COB_dimming.tb$mA <- as.numeric(gsub("^CC.|mA$", "", COB_dimming.tb$spct.idx))
+comment(COB_dimming.tb) <- paste("Constant current dimming. LED current (mA),",
+                                 "photon irradiance (Q_Total), energy irradiance (E_Total),",
+                                 "name of spectrum in 'COB_dimming.mspct' (spct.idx).",
+                                 "LED: Optisolis type NFCWL036B_V3 10W COB", 
+                                 "from Nichia (https://www.nichia.co.jp/)\n", 
+                                 "With LEDiL F15559_MIRELLA-G2-M reflector 25 degrees.")
+COB_dimming.tb
+
+print(ggplot(COB_dimming.tb, aes(mA, Q_Total_rel)) +
+  geom_line())
+
+save(COB_dimming.mspct, COB_dimming.tb, file = "./data/COB-dimming-mspct.rda")